# src/main/java/com/yassine/practice/Authetincation/AuthenticationResponse.java
package com.yassine.practice.Authetincation;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
@Builder
public class AuthenticationResponse {
    private String accessToken;
}