# src/main/java/com/yassine/practice/Email/MailingRespDto.java
package com.yassine.practice.Email;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
@Builder
public class MailingRespDto {
    private String message;
}